refactor api graph tests
